Adds Currents and Examples
Adds the abstract current class as well as two concrete classes, currentWire and currentGaussian, which inherit from the first.
These handle classes provide methods to compute their resulting magnetic field (or flux function), and can be linked together so that a current's children has a fractional value of its parent's current.

A mConf class is also provided to store current handles. It will provide functionalities such as computing the full magnetic field and psi function for a configuration, x-point detection routines, and maybe return analytical expressions for the fields (to be seen later).

mConf will be expanded with relevant properties so it may be used across a toolbox, by passing it to functions for example.