Add Core Detection
The core is now detected accurrately. This method is based on calling vpasolve with initial guess close to the plasma current. The implemented solution requires the guess to be slightly distant from the plasma current (achieved by calling rand).

Future versions could try to get rid of this rand call for mor reliability, or provide input parameters controlling this behavior.