Add Safety Factor Computation (loc and avg)
Local and average computations of the safety factor. Input of parameters is dirty, requires a fix.

The mConf class now has a parameter to indicate on what physical limits a future simulation would be run, without removing generality. This helps, for example, for contour gathering tasks, and might allow to plot the configuration from the class later on.