Fix Gaussian Current Flux Function
Fixes the sign of the exponential integral term in the computation of the flux function for a gaussian current distribution. Integrating the magnetic field to obtain the poloidal flux function, one arrives at

0.5*( log(d2) - Ei(-d2) )

where Ei is the exponential integral, and d2 the squared distance to the current distribution center. MATLAB provides the E1 (E-one) function, expint. The correspondence between these functions can be found to be E1(x) = -Ei(-x).

The sign of the integral was wrong in the previous implementation.