Add Input Parser to main method
The safetyFactor method gets an input parser which will control its subordinate methods. We allow all optional parameters, including
- Number of sampled points
- Target point from the core's center
- Normalizing the output coordinates
- Choice of units (psi or dist) of the output
- A switch to ignore the first point, which might be evaluated to Inf or NaN depending on numerical precision

The choice of a "reasonable" target point still needs to be achieved (a suggestion would be to use a, the plasma width, that would be computed during a magnetic configuration commit).