Add LCFS Detection
Up to now, it was impossible to determine whether a separatrix value of
Psi was the one from a LCFS or not. This can be problematic for a SF+
configuration, for example, where the secondary x-point has lower psi
than the main x-point.

The usual assumption that min(psiSeparatrix)=psiLCFS is then wrong, and
we provide this method for this purpose. It searches the closed contour
with highest psi around the core.