Make Safety Factor and Magnetic Shear Non-Methods
Previously, the functions safetyFactor and magShear were included as
methods of the mConf class. The kind of computation they perform can
however be handled separately to avoid growing the size of the class
beyond reasonable limits. Think of it like in the symbolic toolbox; you
get symbolic functions, but have to call an independent method to
differentiate it.

While the computation of the safety factor runs just fine in the current
version of the code, the magnetic shear will require a small overhaul to
break dependencies from the localSafetyFactor and averageSafetyFactor
methods.