Adds Triangular Configurations
These configuration have just been checked from a geometrical POV. The
safety factor and magnetic shear still need to be adjusted. Positive
configurations are obtained by mirroring the x location of the currents
with respect to the center of the simulated area.

The computation of the various geometrical radiuses was adjusted with a
larger psiLCFS offset.