Add triangular SN configurations
These configurations display only lower triangularity and do not have flux expansion on the top of the domain.

A new function is provided and allows the user to provide a set of mConf objects with varying current properties and generates linear interpolation of the said configurations. This is a one-dimensional parameter scan if you will.