Change mConf Constructor Behavior
Before, when passing currents to the mConf constructor, if an included current's Parent was not found a warning was issued to the user.

We now want to be able to copy mConf object, and the constructor will now issue an error in the previous case. This is because currents do not perform a deep copy of their Parent, such that after copying a mConf, the two objects would still be dependent.

This scenario is of course unwanted.