Allow More Configurations
Until a commit, the mConf class should not worry about missing a plasma current, or if this current's sign is wrong.
Actually, this behavior is already applied when working with symbolic configurations. The only point where an error is issued is when the user tries to commit.
The checkCommit method is modified in consequence to also return an error string and MSGID to explain why a commit is refused.