Commit Options can be Passed to configBrowser
configBrowser can now be called with an argument name - cell array pair, option name 'CommitParams', to provide custom commit options if needed (suppose he needs to adjust 'OffsetScale', this will definitely come in handy).
The input is parsed to avoid empty cells being passed. It also detects whether or not the user is playing with the 'Force' option. In the latter case, when set to false, the 'Re-commit' button is disabled (sadface).